ml-engineer/requirements.txt
google-cloud-bigquery
google-cloud-bigquery-storage
pyarrow
